requirements.txt
pymodbus==3.11.4
PyYAML==6.0.3
